requirements.txt
streamlit
google-generativeai
pymupdf
python-docx
sentence-transformers
chromadb